requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
pydantic
twitchio
jinja2